DAC finally working perfectly

+++ ARMproject/dsping.h
@@ -2,7 +2,7 @@
 #ifndef DSPING_H
 #define	DSPING_H
 
-int get_adc_sample(void);
+uint8_t get_adc_sample(void);
 
 void give_dac_sample(uint8_t sample);
